project/requirements.txt
flask             
flask_sqlalchemy  
pymysql           
python-dotenv     
openai            
rapidfuzz         
requests          
PyPDF2            
numpy
